Fix attachment_url for apps without a root URL
Hello @jnicklas and great job with this gem!

I tried using your gem in a Rails app that does not declare a `root` route in `config/routes.rb` and I found that it does not work
because `attachment_url` is falling back to `root_url`.

If I understand the code correctly, you should fall back to `request.base_url` instead, which will give you the base URL for you app, such as `http://localhost:3000` while you are developing.