Add ActiveRecord test for replacing files
Assignment of the attachment attribute caches the attachment, but
doesn't change any actual columns on the model. Sequel, for example,
won't save the record if no columns were changed, thus when replacing an
attachment on an existing record it won't trigger the `before_save`
callback. So, I thought it's good to have this check for ActiveRecord,
just in case.