Pass :index option through to hidden field
The hidden field and file field need to render with identical `name` attributes
in order for file upload to work properly. The name is generated by Rails based
on certain options, so we need to make sure that both the field and file receive
the same options, at least those options that affect the name.

:index is one of those options. If we call `attachment_field(:image, index: 1)`,
for example, omitting the :index option in the hidden field means we'll render
inputs with different names:

<input type=hidden name=object[image] ...>
<input type=file   name=object[1][image] ...>

This is a bug. This commit fixes this discrepancy by passing the :index option
to the hidden field so that the output is correct:

<input type=hidden name=object[1][image] ...>
<input type=file   name=object[1][image] ...>