Refactor refile.js to be accessible
This PR is a drop in replacement for the previous version of refile.js.
However, it provides a unified API which other JS utilities can access
(a drag and drop library or WYSIWYG editor.

This refactor surfaced a likely bug, where the finishing of the
upload(s) is triggered before the last upload:complete event was
triggered.