Allow assignment of attachment that was previously nil
Currently it's not possible to assign a attachment that was previously
set to nil without recreating the object.

Previously the following code would not work:

```ruby
post.document = nil
post.save
post.document = Refile::DoubleFile.new("hello")
post.save

post.document # => still nil
```

This bug was caused by the `remove` attribute which was set to `true`
when `nil` was assigned but never reset to `false` when a valid
attachment was assigned afterwards.

The previous workaround was to re-create the object eq. calling
something like `post = Post.find(post.id)`